price data type to double

+++ features/merchant/atc_variant/src/main/java/com/tkpd/atcvariant/view/viewmodel/AtcVariantViewModel.kt
@@ -336,7 +336,7 @@ class AtcVariantViewModel @Inject constructor(
                shopIdInt: Int,
                categoryName: String,
                userId: String,
-               shippingMinPrice: Int,
+               shippingMinPrice: Double,
                trackerAttributionPdp: String,
                trackerListNamePdp: String,
                isTokoNow: Boolean) {
