+++ features/merchant/product_detail/src/main/java/com/tokopedia/product/detail/view/widget/FtPDPInstallmentBottomSheet.kt
@@ -37,7 +37,7 @@ class FtPDPInstallmentBottomSheet : BottomSheetDialogFragment() {
 
     private var installmentDataId: String = ""
     private var installmentData: FtInstallmentCalculationDataResponse = FtInstallmentCalculationDataResponse()
-    private var productPrice: Float = 0f
+    private var productPrice: Double = 0.0
     private var isOfficialStore: Boolean = false
 
     private fun getBaseLayoutResourceId(): Int {
@@ -163,7 +163,7 @@ class FtPDPInstallmentBottomSheet : BottomSheetDialogFragment() {
                     FtInstallmentCalculationDataResponse::class.java.simpleName,
                     FtInstallmentCalculationDataResponse::class.java
             ) ?: FtInstallmentCalculationDataResponse()
-            productPrice = it.getFloat(KEY_PDP_PRODUCT_PRICE)
+            productPrice = it.getDouble(KEY_PDP_PRODUCT_PRICE)
             isOfficialStore = it.getBoolean(KEY_PDP_IS_OFFICIAL)
         }
     }
